requirements.txt
# Upgrade pip: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade pip
# Install with: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt

#whell
wheel >= 0.30.0

# MySQL
PyMySQL >= 0.7.11
mysql-replication > 0.18

# ES
elasticsearch >= 6.2.0

